backend/python/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
websockets==11.0.3
httpx==0.25.0
requests==2.31.0
pillow==10.0.1
numpy==1.26.0
pandas==2.1.1 